move config (clap) structs into config.rs
- copied all config structs into config.rs, along with all `use`
statements
- made everything `pub`
- changed main.rs to import these structs
